Fix limit and continue for database listing
There were a few issues related to listing with limits when querying
from a database. First, the predicate limit was incremented after the
criteria was instantiated. The intention seems to be that we only
provide a continue parameter when it is confirmed that there are other
entries to list. Now the predicate limit is incremented prior to the
criteria being instantiated.

When it was determined that a continue parameter should be set, the
above ensured it would happen. However, in order to tell clients that
more items remain, the remaining count should also be set. It seems that
Kubernetes always sets this to 1 for listing. This change does the same.

Lastly, the above two changes alone would cause the client to request
more items, but mink would throw an error complaining about compaction
validation. After this change, we will ignore compaction validation when
utilizing the continue parameter, similar to the implementation for
watch initialization.

Signed-off-by: Donnie Adams <donnie@acorn.io>