Stop leaking goroutines if compaction check fails
It was discovered in one of our services that there was a memory leak.
An inspection of the stack indicated that there were 25729 goroutines
and 25571 were stuck on line 24 of pkg/channel/concat.go. The only way
to get stuck on this line is if neither of the channels passed to Concat
ever close.

Upon examination of the Watch that calls channel.Concat, the only way
that neither channel would be closed is if neither of the subsequent
goroutines were started. Hence, the validateCriteria check failed.

These changes ensure that the channels are closed in this case.
Additionally, it is ensured that the subsequent created goroutine
returns when the merged channel is closed.

Signed-off-by: Donnie Adams <donnie@acorn.io>