Add partition ID for a multi-tenant approach to databases
A partition ID is added to the database strategy to allow for
multi-tenant storage in the database layer. A parameter is added to the
database factory to make the partition ID required for all requests. If
the partition ID is required and none is put on the context, then an
internal server error is returned.

Signed-off-by: Donnie Adams <donnie@acorn.io>